# intent-layer-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Intent Layer API
  description: |
    Intent Layer backend (MVP 1.5). Parse natural-language prompts to structured SearchIntent,
    fetch candidate spaces from DB, and return LLM-ranked recommendations.
    - **Parse** — prompt → SearchIntent JSON
    - **Search** — intent JSON → ranked results
    - **Search by prompt** — prompt → intent + ranked results (all-in-one)
  version: 1.0.0
  contact:
    name: Intent Layer API

servers:
  - url: http://localhost:3001
    description: Development server
  - url: http://144.172.91.222:20001
    description: Test server
  - url: https://driverpacks.space
    description: Online server

paths:
  /api/intent/parse:
    post:
      tags:
        - Intent
      summary: Parse prompt to structured intent
      description: Accepts a natural-language or typed prompt and returns a structured SearchIntent (date, duration, noOfGuests, location, categoryIds).
      operationId: parseIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseRequest"
      responses:
        "200":
          description: Structured intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseResponse"
        "400":
          description: Bad request (e.g. missing prompt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Intent extraction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorWithDetails"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/intent/search:
    post:
      tags:
        - Search
      summary: Search by intent
      description: Input intent JSON. Fetches candidates from DB and returns LLM-ranked recommendations with scores and reasons.
      operationId: searchByIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Ranked recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Missing intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Search failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/intent/search-by-prompt:
    post:
      tags:
        - Search
      summary: Search by prompt (all-in-one)
      description: Input raw prompt. Runs parse → fetch candidates → LLM rank. Returns intent, repair, and ranked recommendations.
      operationId: searchByPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchByPromptRequest"
      responses:
        "200":
          description: Intent + ranked recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchByPromptResponse"
        "400":
          description: Missing prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Intent extraction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorWithDetails"
        "500":
          description: Search failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns service status.
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: intent-layer-backend

components:
  schemas:
    ParseRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          description: User's typed or spoken text
          example: "Breakfast for 2 in London next Tuesday 10am for 2 hours"
        context:
          type: object
          properties:
            timezone:
              type: string
              example: "Europe/London"
            lastLocation:
              type: object
              properties:
                lat:
                  type: number
                  example: 51.576525
                lng:
                  type: number
                  example: -0.153765
                address:
                  type: string
                  example: "14 Grange Road, London N6 4DG"

    ParseResponse:
      type: object
      required:
        - intent
      properties:
        intent:
          $ref: "#/components/schemas/SearchIntent"
        repair:
          type: object
          properties:
            applied:
              type: boolean
            changes:
              type: array
              items:
                type: string

    SearchRequest:
      type: object
      required:
        - intent
      properties:
        intent:
          $ref: "#/components/schemas/SearchIntent"
        sessionId:
          type: string
          description: Clerk session ID (forwarded when using allspaces HTTP fallback)
        categoryIds:
          type: array
          items:
            type: string
          description: Optional category UUIDs to filter

    SearchResponse:
      type: object
      properties:
        intent:
          $ref: "#/components/schemas/SearchIntent"
        repair:
          type: object
          description: Present when intent was normalized server-side (e.g. legacy body).
          properties:
            applied:
              type: boolean
            changes:
              type: array
              items:
                type: string
        summary:
          type: string
          example: "3 Italian restaurants in London for your Friday dinner."
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        candidatesCount:
          type: integer
          description: Total candidates found before ranking
        noMatchMessage:
          type: string
          nullable: true
          description: Message when no suitable candidates

    SearchByPromptRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          example: "Italian restaurant in London for 4 people this Friday at 7pm"
        context:
          type: object
          properties:
            timezone:
              type: string
              example: "Europe/London"
            lastLocation:
              type: object
              properties:
                lat:
                  type: number
                  example: 51.576525
                lng:
                  type: number
                  example: -0.153765
                address:
                  type: string
                  example: "14 Grange Road, London N6 4DG"
        sessionId:
          type: string

    SearchByPromptResponse:
      type: object
      properties:
        intent:
          $ref: "#/components/schemas/SearchIntent"
        repair:
          type: object
          properties:
            applied:
              type: boolean
            changes:
              type: array
              items:
                type: string
        summary:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        candidatesCount:
          type: integer
        noMatchMessage:
          type: string
          nullable: true

    SearchIntent:
      type: object
      description: |
        Search-ready intent from `/parse` or manual construction. Supports legacy flat shape
        (`date` as ISO string, top-level `duration`, `noOfGuests`) — the server normalizes to this model.
      properties:
        searchType:
          type: string
          enum: [DISCOVERY, BOOKING]
        entityTypes:
          type: array
          items:
            type: string
            enum: [PROFILE, SERVICE]
        categoryType:
          type: string
          nullable: true
          description: Prisma CategoryType
          enum:
            - WORKSPACE
            - RELAXATION
            - DINING
            - SLEEP
            - TRAVEL
            - POLICIES
            - SPECIAL_OFFERS
            - BEACH_CLUB
        categoryIds:
          type: array
          items:
            type: string
          description: Category UUIDs or semantic IDs (e.g. cat-dining-001)
        serviceCategoryIds:
          type: array
          items:
            type: string
        rawQuery:
          type: string
        normalizedQuery:
          type: string
        inferredDescription:
          type: string
        date:
          type: object
          properties:
            checkIn:
              type: string
              format: date-time
              nullable: true
            checkOut:
              type: string
              format: date-time
              nullable: true
            duration:
              type: integer
              nullable: true
            timeText:
              type: string
              nullable: true
            serviceTime:
              type: string
              format: date-time
              nullable: true
            timeOfDay:
              type: string
              nullable: true
            isFlexible:
              type: boolean
        guests:
          type: object
          properties:
            noOfGuests:
              type: integer
            rooms:
              type: integer
              nullable: true
            adults:
              type: integer
              nullable: true
            children:
              type: integer
              nullable: true
        location:
          type: object
          properties:
            queryText:
              type: string
              nullable: true
            address:
              type: string
              nullable: true
            lat:
              type: number
              nullable: true
            lng:
              type: number
              nullable: true
            radiusKm:
              type: number
              nullable: true
            nearUser:
              type: boolean
        userLocation:
          type: object
          nullable: true
          properties:
            address:
              type: string
            lat:
              type: number
            lng:
              type: number
        price:
          type: object
          properties:
            min:
              type: number
              nullable: true
            max:
              type: number
              nullable: true
            currency:
              type: string
              nullable: true
            isBudgetIntent:
              type: boolean
            isLuxuryIntent:
              type: boolean
        rating:
          type: object
          properties:
            minAverageRating:
              type: number
              nullable: true
            minReviewCount:
              type: integer
              nullable: true
        facilities:
          type: object
          properties:
            required:
              type: array
              items:
                type: string
            preferred:
              type: array
              items:
                type: string
            excluded:
              type: array
              items:
                type: string
        serviceFilters:
          type: object
          properties:
            minSpend:
              type: number
              nullable: true
            durationLabel:
              type: string
              nullable: true
            timeOfDay:
              type: string
              nullable: true
        profileFilters:
          type: object
          properties:
            source:
              type: array
              items:
                type: string
            status:
              type: array
              items:
                type: string
        semantic:
          type: object
          properties:
            mustTerms:
              type: array
              items:
                type: string
            shouldTerms:
              type: array
              items:
                type: string
            excludeTerms:
              type: array
              items:
                type: string
        serviceLabels:
          type: array
          description: Customer catalog labels (exact strings). Used with category inference and ranking.
          items:
            $ref: "#/components/schemas/CustomerServiceLabel"
        sort:
          type: object
          properties:
            by:
              type: string
              enum: [RELEVANCE, DISTANCE_THEN_RELEVANCE, RATING, PRICE]
            direction:
              type: string
              enum: [ASC, DESC]
        confidence:
          type: number
          minimum: 0
          maximum: 1

    CustomerServiceLabel:
      type: string
      enum:
        - Workspace
        - Dining Services
        - Relax & Leisure
        - Beach Club
        - Sleep Facilities
        - Travel
        - Policies
        - Special Offers
        - Lobby Workspace
        - Business Room
        - Conference Room
        - Board Room
        - IT Facilities
        - Breakfast
        - Lunch
        - Dinner
        - Coffee / Tea
        - Spa
        - Massage
        - Gym
        - Swimming
        - Yoga
        - Pilates
        - Classes
        - Beach Bed
        - Dining Options
        - Champagne
        - Room (Full Day)
        - Half Day
        - Hourly (min. 4 hrs)
        - Uber
        - Taxi
        - Bike
        - Flight
        - Non-Smoking
        - Smoking
        - Free Coffee
        - Complimentary Champagne
        - Service Discounts

    Recommendation:
      type: object
      properties:
        rank:
          type: integer
          description: 1-based rank
        profileId:
          type: string
          format: uuid
        profileName:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1
          description: Match score 0-1
        reasons:
          type: array
          items:
            type: string
          description: Why this space was recommended

    Error:
      type: object
      properties:
        message:
          type: string

    ErrorWithDetails:
      type: object
      properties:
        message:
          type: string
        details:
          type: string